refactor: change directory structure and apply to files (#18)
