+++ strawberry/src/core/design_system/border.tsx
