+++ strawberry/src/main.tsx
@@ -1,7 +1,7 @@
 import App from "./App.tsx";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import theme from "./modules/core/design_system/theme.tsx";
+import theme from "./core/design_system/theme.tsx";
 import { ThemeProvider } from "styled-components";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
